Create the one where we help the police
A lot of police investigations start off by examining where the potential suspects was at the time of the crime. This is often done by asking phone companies to provide lists of all mobile phones that was in the area where the crime was committed around the time when it was committed. If there is only one crime scene this is not a big problem but if there are several crime scenes then the task of cross referencing the lists gets quite tedious. To combat this the local police department has decided to automate the process by creating a script that automatically does most of the work for them.

The script works as follows:

(Step 1) Start the script
(Step 2) Provide one or more paths to files with phone numbers
(Step 3) The script finds out what phone numbers occur in all files
(Step 4) The script looks up who owns the phone numbers found
(Step 5) The script displays the found names